+++ HappyFarmProject/HappyFarmProjectWebAdmin/Controllers/Super Admin/SuperAdminSubDistrictController.cs
@@ -314,7 +314,7 @@ namespace HappyFarmProjectWebAdmin.Controllers.Super_Admin
 
                 ModifiedBy = (int)Session["UserId"],
                 Name = subDistrict.Name,
-                ShippingCharges = subDistrict.ShippingCharges,
+                ShippingCharges = (long)subDistrict.ShippingCharges,
                 RegionId = subDistrict.RegionId
             };
 
@@ -382,7 +382,7 @@ namespace HappyFarmProjectWebAdmin.Controllers.Super_Admin
             {
                 ModifiedBy = (int)Session["UserId"],
                 Name = subDistrict.Name,
-                ShippingCharges = subDistrict.ShippingCharges,
+                ShippingCharges = (long)subDistrict.ShippingCharges,
                 RegionId = subDistrict.RegionId
             };
